Move google search logic inside a class
