+++ IR/src/google-search/index.ts
@@ -1,5 +1,7 @@
-import GoogleSearchRouter from './google-search';
+import GoogleSearchRouter from './google-search.rest';
+import { GoogleSearch } from './google-search.module'
 
 export {
-    GoogleSearchRouter
+    GoogleSearchRouter,
+    GoogleSearch
 }
